Restructured tests as well
