+++ src/algorithm/algorithm_functions.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "graph_functions.h"
 #include <math.h>
 #define NODE_ATTRACTION_FACTOR 0.04
