+++ src/dataparser/csv_parser.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <stdio.h>
 #include <stdlib.h>
 #include "graph_functions.h"
