+++ src/sample_network/CMakeLists.txt
@@ -3,3 +3,5 @@ add_library(sample_network
 )
 
 target_include_directories(sample_network PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_link_libraries(dataparser PUBLIC graph)
