+++ src/sample_network/sample_network.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "graph_functions.h"
 #define TEST_NETWORK_NODE_AMOUNT 10
