+++ test/dataparser/test_csv_parser.c
@@ -1,3 +1,21 @@
-int main(){
-    
+#include "csv_parser.h"
+#include "unity.h"
+
+void setUp(void) {
+    // set stuff up here
+}
+void tearDown(void) {
+    // clean stuff up here
+}
+void test_function_should_doBlahAndBlah(void) {
+    //test stuff
+}
+void test_function_should_doAlsoDoBlah(void) {
+    //more test stuff
+}
+// not needed when using generate_test_runner.rb
+int main(void) {
+    UNITY_BEGIN();
+    RUN_TEST(test_function_should_doBlahAndBlah);
+    RUN_TEST(test_function_should_doAlsoDoBlah);
 }
